+++ common/src/main/java/de/cech12/usefulhats/item/ILivingJumpListener.java
@@ -1,0 +1,10 @@
+package de.cech12.usefulhats.item;
+
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+
+public interface ILivingJumpListener {
+
+    void onLivingJumpEvent(LivingEntity jumpingEntity, ItemStack headSlotItemStack);
+
+}