+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.20.4-5.1.2.0] - 2024-05-11
+### Changed
+- Bunny Ears only get damage when jumping or when food is eaten
+- reduced Bunny Ears default durability from 600 to 450
+
 ## [1.20.4-5.1.1.1] - 2024-05-11
 ### Fixed
 - Respiration effect of Aquanaut Helmet was removed when getting underwater (all loaders)